// source/types/offersTypes.ts
export interface DiscountUsageParameters {
  /** Reference to reserving parameters with price information. */
  reserving?: PricedUsageParameter[];

  /** Reference to entitlement given parameters with price information. */
  entitlementGiven?: PricedUsageParameter[];
}
// ...
export interface SalesPackageConfiguration {
  /**
   * The ID of the object on the format < context : type : instance name >
   * @example NSB:SalesPackage:Ordinary
   */
  id: string;

  /**
   * The version of the object
   * @example NSB:Version:V1
   */
  version: string;

  /** The name of the object. */
  name: TextInLanguage[];

  /** The description of the object. */
  description: TextInLanguage[];

  /** The price(s) of the sub configurations below this node. Multiple prices is supported in case of prices in multiple currencies. */
  prices: Price[];

  /** The price contribution shows how much this node contributes on the total price. */
  priceContributions: Price[];

  /** Container object for other properties. */
  properties?: Properties;

  /** Reference to different UsageParameters. */
  parameters?: UsageParameters;

  /** Reference to the legs this configuration is for. */
  serviceJourneyIds: string[];

  /** Reference to the travelscope this configuration is for. */
  travelScopeIds?: string[];

  /** Reference to the legs this configuration is for. */
  datedServiceJourneys?: DatedServiceJourney[];

  /** References to different distribution parameters. */
  distributions: DistributionParameters[];

  /** Reference to a Fare Product and its configuration. */
  fareProducts: FareProductConfiguration[];
}
// ...
export interface OrganisationConfig {
  /** The netexId for the organisation. */
  id: string;

  /**
   * The privateCode for the organisation, used to integrate with the organisation register.
   * @format int64
   */
  privateCode?: number;
}
// ...
export interface LineValidity {
  /** Valid on lines */
  validOn?: string[];

  /** Not valid on lines */
  notValidOn?: string[];
}
// ...
export interface Tag {
  /** An identifier for the TAG. */
  name: string;

  /**
   * How many of this offer should be bought to cover the TRAVELERs
   * @format int32
   */
  nrToBuy: number;
}
// ...
export interface ConditionSummary {
  /**
   * Whether the ticket is refundable.
   * @example true
   */
  isRefundable?: boolean;

  /**
   * Whether the ticket is exchangeable.
   * @example true
   */
  isExchangeable?: boolean;
}
// ...
export interface PointToPointValidity {
  /**
   * Valid from place
   * @example NSR:StopPlace:398
   */
  fromPlace?: string;

  /**
   * Valid to place
   * @example NSR:StopPlace:440
   */
  toPlace?: string;
}
// ...
export interface PricedResourceRef {
  /** ID of the resource, e.g. an ID of a 'TripPattern'. */
  id: string;

  /** Name of the resource. */
  resourceType: string;
}
// ...
export interface DatedServiceJourney {
  /**
   * A unique departure ID.
   * @example ENT:DatedServiceJourney:151530
   */
  datedServiceJourneyId: string;

  /**
   * A reference to the original dated service journey ID.
   * @example ENT:DatedServiceJourney:151530
   */
  originalDatedServiceJourneyId: string;
}
// ...
export interface ValidableElementConfiguration {
  /**
   * The ID of the object on the format < context : type : instance name >
   * @example NSB:ValidableElement:Ordinary
   */
  id: string;

  /**
   * The version of the object
   * @example NSB:Version:V1
   */
  version: string;

  /** The name of the object. */
  name: TextInLanguage[];

  /** The description of the object. */
  description: TextInLanguage[];

  /** The price(s) of the sub configurations below this node. Multiple prices is supported in case of prices in multiple currencies. */
  prices: Price[];

  /** The price contribution shows how much this node contributes on the total price. */
  priceContributions: Price[];

  /** Container object for other properties. */
  properties?: Properties;

  /** Reference to different UsageParameters. */
  parameters?: UsageParameters;

  /** Reference to the legs this configuration is for. */
  serviceJourneyIds: string[];

  /** Reference to the legs this configuration is for. */
  datedServiceJourneys?: DatedServiceJourney[];

  /** Reference to a Fare Structure Element and its configuration. */
  fareStructureElements: FareStructureElementConfiguration[];

  /**
   * How many repeatable validable elements
   * @format int32
   * @example 10
   */
  occurs?: number;
}

/**
 * Text in a specific language.
 */
export interface TextInLanguage {
  /** Language code. */
  lang?: string;

  /** The text. */
  value?: string;
}
// ...
export interface Offer {
  /**
   * The ID of the OFFER.
   * @format uuid
   */
  id: string;

  /** The requested distribution channel */
  requestedByDistributionChannel?: string;

  /** A reference to the resource that is priced in this offer, e.g. a specific 'TripPattern'. */
  pricedResource?: PricedResourceRef;

  /** Reference to a Sales Package and its configuration. */
  salesPackageConfig: SalesPackageConfiguration;

  /** Traveler limitations that applies to the SALES PACKAGE CONFIGURATION. */
  travelerMapping: TravelerValidityGroup[];

  /** All OFFERs within the same TAG are covering as many as possible SERVICE JOURNEYs for all TRAVELERs. */
  tags: Tag[];
}

/**
 * Price with VAT and currency information
 */
export interface Price {
  /** The amount. */
  amount: string;

  /** The VAT group. */
  vatGroup?: string;

  /**
   * The currency of the amount.
   * @example NOK
   */
  currency: string;
}

/**
 * Reference to different UsageParameters.
 */
export interface UsageParameters {
  /** A reference to a user profile and price data. */
  userProfile?: PricedUserProfile;

  /** Reference to a refund parameter with price information. */
  refunds: PricedUsageParameter[];

  /** A reference to a user parameter and price data. */
  usageValidityPeriod?: PricedUsageParameter;

  /** Reference to luggage allowance parameters with price information. */
  luggageAllowances: PricedUsageParameter[];

  /** Reference to exchanging parameters with price information. */
  exchanging: PricedUsageParameter[];

  /** Reference to reserving parameters with price information. */
  reserving: PricedUsageParameter[];

  /** Reference to replacing parameters with price information. */
  replacing: PricedUsageParameter[];

  /** Reference to purchase window parameters with price information. */
  purchaseWindow: PricedUsageParameter[];

  /** Reference to entitlement given parameters with price information. */
  entitlementGiven: PricedUsageParameter[];

  /** Reference to round trip parameters with price information */
  roundTrip: PricedUsageParameter[];

  /** A reference to a group with price data. */
  group?: PricedGroup;

  /** Reference to transferability parameters with price information. */
  transferability?: PricedUsageParameter[];
}

/**
 * Reference to a Fare Product and its configuration.
 */
export interface FareProductConfiguration {
  /**
   * The ID of the object on the format < context : type : instance name >
   * @example NSB:PreassignedFareProduct:Ordinary
   */
  id: string;

  /** Used for selecting a unique element in the configurations */
  selectableId: string;

  /**
   * The version of the object
   * @example NSB:Version:V1
   */
  version: string;

  /** The name of the object. */
  name: TextInLanguage[];

  /** The description of the object. */
  description: TextInLanguage[];

  /** Organisation that owns the given product. */
  organisation: OrganisationConfig;

  /** The price(s) of the sub configurations below this node. Multiple prices is supported in case of prices in multiple currencies. */
  prices: Price[];

  /** The price contribution shows how much this node contributes on the total price. */
  priceContributions: Price[];

  /** Container object for other properties. */
  properties?: Properties;

  /** Reference to different UsageParameters. */
  parameters?: UsageParameters;

  /** Reference to the legs this configuration is for. */
  serviceJourneyIds: string[];

  /** Reference to the legs this configuration is for. */
  datedServiceJourneys?: DatedServiceJourney[];

  /** True if this configuration is optional to buy, false if it's required. */
  optional: boolean;

  /** Reference to a Validable Element and its configuration. */
  validableElements: ValidableElementConfiguration[];

  /** A classification of this fare product */
  typeOfFareProduct?: string;

  /** A summary of the properties of a fare product or package that can be used to generate passenger information. */
  conditionSummary?: ConditionSummary;

  /** A sale discount right that has been applied to the fare product. */
  discountRight?: DiscountRight;

  /** The id of the entitlement product that allowed the sale discount right to be used. Only used if the fare product configuration contains a sale discount right, and the traveler had an entitlement product. */
  originatingFromProductId?: string;

  /** Start and stop places where the product is valid */
  pointToPointValidity?: PointToPointValidity;

  /** Zones where the product is valid */
  zonalValidity?: ZonalValidity;

  /** Lines where the product is valid */
  lineValidity?: LineValidity;

  /**
   * Amount of priced unit
   * @format int32
   */
  amount?: number;

  /** Type of priced unit */
  productType?:
    | "TRIP_CARNET"
    | "PASS_CARNET"
    | "UNIT_COUPONS"
    | "STORED_VALUE"
    | "OTHER";
}
// ...
export interface ZonalValidity {
  /**
   * Valid in zones
   * @example ['RUT:TariffZone:1']
   */
  zones?: string[];
}
// ...
export interface DiscountRight {
  /**
   * The ID of the object on the format < context : type : instance name >
   * @example SJN:SaleDiscountRight:Premium
   */
  id: string;

  /** Used for selecting a unique element in the configurations */
  selectableId: string;

  /**
   * The version of the object
   * @example SJN:Version:SDR-Premium-1
   */
  version: string;

  /** The name of the object. */
  name?: TextInLanguage[];

  /** The description of the object. */
  description?: TextInLanguage[];

  /** Organisation that owns the given product. */
  organisation: OrganisationConfig;

  /** The price(s) of the sub configurations below this node. Multiple prices is supported in case of prices in multiple currencies. */
  prices: Price[];

  /** The price contribution shows how much this node contributes on the total price. */
  priceContributions: Price[];

  /** Reference to different UsageParameters. */
  parameters?: DiscountUsageParameters;

  /** True if this configuration is optional to buy, false if it's required. */
  optional: boolean;

  /** The id of the entitlement product that allowed the sale discount right to be used. Only used if the fare product configuration contains a sale discount right, and the traveler had an entitlement product. */
  originatingFromProductId?: string;
}

/**
 * Container object for other properties.
 */
export interface Properties {
  /** FareClasses. Enum as string with the following values: FIRST_CLASS, SECOND_CLASS, THIRD_CLASS, ECONOMY_CLASS, BUSINESS_CLASS, TURISTA, PREFERENTE, STANDARD_CLASS, PREMIUM_CLASS, ANY, UNKNOWN */
  fareClasses?: (
    | "FIRST_CLASS"
    | "SECOND_CLASS"
    | "THIRD_CLASS"
    | "ECONOMY_CLASS"
    | "BUSINESS_CLASS"
    | "TURISTA"
    | "PREFERENTE"
    | "STANDARD_CLASS"
    | "PREMIUM_CLASS"
    | "ANY"
    | "UNKNOWN"
  )[];

  /** Accommodation. Enum as string with the following values: SLEEPER, SEATING, DOUBLE_SLEEPER, SINGLE_SLEEPER, SPECIAL_SLEEPER, COUCHETTE, SINGLE_COUCHETTE, DOUBLE_COUCHETTE, RECLINING_SEAT, BABY_COMPARTMENT, FAMILY_COMPARTMENT, PANORAMA_COACH, PULLMAN_COACH, STANDING */
  accommodations?: (
    | "SLEEPER"
    | "SEATING"
    | "DOUBLE_SLEEPER"
    | "SINGLE_SLEEPER"
    | "SPECIAL_SLEEPER"
    | "COUCHETTE"
    | "SINGLE_COUCHETTE"
    | "DOUBLE_COUCHETTE"
    | "RECLINING_SEAT"
    | "BABY_COMPARTMENT"
    | "FAMILY_COMPARTMENT"
    | "PANORAMA_COACH"
    | "PULLMAN_COACH"
    | "STANDING"
  )[];
}